remove deprecated IndexAnalysisSettings (version 1) fixes #2290
