+++ opengrok-indexer/src/main/java/org/opengrok/indexer/web/SearchHelper.java
@@ -62,7 +62,7 @@ import org.opengrok.indexer.analysis.Definitions;
 import org.opengrok.indexer.configuration.Project;
 import org.opengrok.indexer.configuration.RuntimeEnvironment;
 import org.opengrok.indexer.configuration.SuperIndexSearcher;
-import org.opengrok.indexer.index.IndexAnalysisSettings2;
+import org.opengrok.indexer.index.IndexAnalysisSettings;
 import org.opengrok.indexer.index.IndexAnalysisSettingsAccessor;
 import org.opengrok.indexer.index.IndexDatabase;
 import org.opengrok.indexer.logger.LoggerFactory;
@@ -217,7 +217,7 @@ public class SearchHelper {
     /**
      * Key is Project name or empty string for null Project
      */
-    private Map<String, IndexAnalysisSettings2> mappedAnalysisSettings;
+    private Map<String, IndexAnalysisSettings> mappedAnalysisSettings;
 
     /**
      * User readable description for file types. Only those listed in
@@ -621,7 +621,7 @@ public class SearchHelper {
      */
     public int getTabSize(Project proj) throws IOException {
         String projectName = proj != null ? proj.getName() : null;
-        IndexAnalysisSettings2 settings = getSettings(projectName);
+        IndexAnalysisSettings settings = getSettings(projectName);
         int tabSize;
         if (settings != null && settings.getTabSize() != null) {
             tabSize = settings.getTabSize();
@@ -639,12 +639,12 @@ public class SearchHelper {
      * @return a defined instance or {@code null} if none is found
      * @throws IOException if an I/O error occurs querying the active reader
      */
-    public IndexAnalysisSettings2 getSettings(String projectName)
+    public IndexAnalysisSettings getSettings(String projectName)
             throws IOException {
         if (mappedAnalysisSettings == null) {
             IndexAnalysisSettingsAccessor dao =
                 new IndexAnalysisSettingsAccessor();
-            IndexAnalysisSettings2[] setts = dao.read(reader, Short.MAX_VALUE);
+            IndexAnalysisSettings[] setts = dao.read(reader, Short.MAX_VALUE);
             mappedAnalysisSettings = map(setts);
         }
 
@@ -652,12 +652,12 @@ public class SearchHelper {
         return mappedAnalysisSettings.get(k);
     }
 
-    private Map<String, IndexAnalysisSettings2> map(
-        IndexAnalysisSettings2[] setts) {
+    private Map<String, IndexAnalysisSettings> map(
+        IndexAnalysisSettings[] setts) {
 
-        Map<String, IndexAnalysisSettings2> res = new HashMap<>();
+        Map<String, IndexAnalysisSettings> res = new HashMap<>();
         for (int i = 0; i < setts.length; ++i) {
-            IndexAnalysisSettings2 settings = setts[i];
+            IndexAnalysisSettings settings = setts[i];
             String k = settings.getProjectName() != null ?
                 settings.getProjectName() : "";
             res.put(k, settings);
